Clean model update calculation
