+++ src/language-server/task-list/semantic/task-list-identity-reconciler.ts
@@ -3,7 +3,7 @@ import { AstRootNode } from '../../../langium-model-server/semantic/model'
 import * as src from '../../../langium-model-server/lms/model'
 import type * as ast from '../../generated/ast'
 import type * as source from '../lms/model'
-import type { TaskListModelUpdateManager } from '../lms/task-list-source-update-manager'
+import type { TaskListModelUpdateCalculators } from '../lms/task-list-source-update-calculation'
 import type { TaskListServices } from '../task-list-module'
 import { type TaskListDocument } from '../workspace/documents'
 import { Model } from './task-list-identity'
@@ -11,11 +11,11 @@ import type { TaskListIdentityManager } from './task-list-identity-manager'
 
 export class TaskListIdentityReconciler implements IdentityReconciler<source.Model, TaskListDocument>{
     private identityManager: TaskListIdentityManager
-    private modelUpdateManager: TaskListModelUpdateManager
+    private modelUpdateCalculators: TaskListModelUpdateCalculators
 
     public constructor(services: TaskListServices) {
         this.identityManager = services.semantic.IdentityManager
-        this.modelUpdateManager = services.lms.ModelUpdateManager
+        this.modelUpdateCalculators = services.lms.ModelUpdateCalculators
     }
 
     /* NOTE: So, the problem can be characterized as following:
@@ -39,13 +39,13 @@ export class TaskListIdentityReconciler implements IdentityReconciler<source.Mod
     private reconcileTasks(document: TaskListDocument, update: src.Update<source.Model>) {
 
         const identityIndex = this.identityManager.getIdentityIndex(document)
-        const updateCalculator = this.modelUpdateManager.getUpdateCalculator(document)
+        const updateCalculator = this.modelUpdateCalculators.getOrCreateCalculator(document)
         const semanticDomain = document.semanticDomain!
         // NOTE: Here I am expressing an idea, that perhaps I will have to have some sort of nested model indices,
         // which would make it generally necessary to pass the parent model into the semantic domain when requesting some (valid/identified) models
         const astModel: ast.Model = document.parseResult.value
 
-        // TODO: Suggest AstRootNode as a specific interface for Langium
+        // TODO: Suggest AstRootNode as a specific interface in Langium library
         const rootNode = AstRootNode.create(astModel)
         if (!rootNode) {
             throw new Error('Expected Model to be a root node, but somehow it was not!. Model: ' + astModel)
@@ -77,7 +77,7 @@ export class TaskListIdentityReconciler implements IdentityReconciler<source.Mod
     private reconcileTransitions(document: TaskListDocument, update: src.Update<source.Model>) {
 
         const identityIndex = this.identityManager.getIdentityIndex(document)
-        const updateCalculator = this.modelUpdateManager.getUpdateCalculator(document)
+        const updateCalculator = this.modelUpdateCalculators.getOrCreateCalculator(document)
         const semanticDomain = document.semanticDomain!
 
         const existingUnmappedTransitions = identityIndex.transitionsByName
